remove: tmdb keywords
This extra data isn't useful for anything except being spoiled about the movie/tv. The only usecase its had in the past was searching for anime, but we have a language filter for search now, and filtering by `ja` and `animation` seems to be more reliable than searching for the `anime` keyword.