refactor: use laravel-data for typed form requests and dtos
The developer experience is much improved with the typed form requests, and matches behaviour common in other type-safe languages. It allows accessing properties directly off the dto that came from the form request with ide autocomplete, and it validates that the properties exist and are not null. It might also be the solution to fixing the livewire type errors where a user supplies a string to an int variable as well.